+++ cypress/e2e/fundamentalTypes/all.cy.js
@@ -9,6 +9,7 @@ import './string.cy.js';
 import './regexp.cy.js';
 import './array.cy.js';
 import './arrayReference.cy.js';
+import './error.cy.js';
 
 import './BooleanObject.cy.js';
 import './NumberObject.cy.js';